requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
black==23.10.1
blessed==1.20.0
bpython==0.24
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
curtsies==0.4.2
cwcwidth==0.1.9
decorator==5.1.1
Django==4.2.6
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
executing==2.0.0
Faker==19.12.0
frozenlist==1.4.0
greenlet==3.0.0
idna==3.4
install==1.3.5
ipython==8.16.1
jedi==0.19.1
magic-filter==1.0.12
matplotlib-inline==0.1.6
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.11.0
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.3.0
pydantic_core==2.6.3
Pygments==2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyxdg==0.28
requests==2.31.0
rest-framework-simplejwt==0.0.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.12.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.8
yarl==1.9.2
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
drf-spectacular==0.26.5
stripe==7.1.0
Pillow==10.1.0
